refactor(psxed-ui): extract gizmo value types and helpers into gizmo module
Move PrimitiveGizmoAxis, NodeGizmoPlane, NodeGizmoHandle and the gizmo
geometry/color helpers out of the psxed-ui monolith into a dedicated
gizmo module. Pure code movement; no behaviour change.
